+++ docs/source/changelog.rst
@@ -14,8 +14,12 @@ History
 Version 0.2.26 (2022-)
 ---------------------------
 
-* Work on ``GeneticProgram`` class.
+* Added tree generation function ``random_generate`` to genetic_programs.
 * Renamed ``natural_selection.genetic_programs.functions `` to ``natural_selection.genetic_programs.node_operators ``.
+* Added ``is_empty`` to Node class.
+* Added more history "stamps" to Island for alien spawn, migrant import.
+* General work on ``GeneticProgram`` class.
+
 
 Version 0.2.25 (2022-06-20)
 ---------------------------